requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
argparse==1.4.0
astroid==1.5.3
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
caniusepython3==7.1.0
certifi==2019.6.16
chardet==3.0.4
click-log==0.3.2
click==7.0
code-annotations==0.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.6.0
diff-cover==2.3.0
distlib==0.2.9.post0
django-cors-headers==3.0.2
django-debug-toolbar==1.11
django-dynamic-fixture==2.0.0
django-extensions==2.1.9
django-modelcluster==3.1
django-nose==1.4.6
django-rest-swagger==2.2.0
django-storages==1.7.1
django-taggit==0.22.2
django-treebeard==4.3
django-waffle==0.17.0
django==1.11.22
djangorestframework-jwt==1.11.0
djangorestframework==3.6.4
edx-auth-backends==2.0.1
edx-django-release-util==0.3.1
edx-django-utils==1.0.5
edx-drf-extensions==2.3.1
edx-i18n-tools==0.4.8
edx-lint==1.3.0
edx-opaque-keys==1.0.1
edx-rest-api-client==1.9.2
factory-boy==2.12.0
faker==1.0.7
future==0.17.1
html5lib==0.999999999
idna==2.8
importlib-metadata==0.18  # via path.py
inflect==2.1.0            # via jinja2-pluralize
isort==4.3.21
itypes==1.1.0
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10.1
lazy-object-proxy==1.4.1
markupsafe==1.1.1
mccabe==0.6.1
mock==3.0.5
mysqlclient==1.4.2.post1
newrelic==4.20.1.121
nose-ignore-docstring==0.2
nose==1.3.7
oauthlib==3.0.2
openapi-codec==1.3.2
packaging==19.0
path.py==12.0.1           # via edx-i18n-tools
pbr==5.4.0
pillow==6.1.0
pip-tools==3.8.0
polib==1.1.0              # via edx-i18n-tools
psutil==1.2.1
pycodestyle==2.5.0
pycryptodomex==3.8.2
pydocstyle==4.0.0
pygments==2.4.2           # via diff-cover
pyjwkest==1.3.2
pyjwt==1.7.1
pylint-celery==0.3
pylint-django==0.7.2
pylint-plugin-utils==0.5
pylint==1.7.6
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
pyyaml==5.1.1
requests-oauthlib==1.2.0
requests==2.22.0
rest-condition==1.0.3
semantic-version==2.6.0
simplejson==3.16.0
six==1.12.0
slumber==0.7.1
snowballstemmer==1.9.0
social-auth-app-django==1.2.0
social-auth-core[openidconnect]==1.7.0
sqlparse==0.3.0           # via django-debug-toolbar
stevedore==1.30.1
text-unidecode==1.2
unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
wagtail==1.13.4
webencodings==0.5.1
willow==1.0
wrapt==1.11.2
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via caniusepython3, html5lib
